+++ engine/runtime/src/main/java/org/enso/interpreter/runtime/Context.java
@@ -4,35 +4,34 @@ import com.oracle.truffle.api.CompilerDirectives.CompilationFinal;
 import com.oracle.truffle.api.TruffleFile;
 import com.oracle.truffle.api.TruffleLanguage;
 import com.oracle.truffle.api.TruffleLanguage.Env;
+import com.oracle.truffle.api.TruffleLogger;
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.PrintStream;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 import org.enso.compiler.Compiler;
 import org.enso.compiler.PackageRepository;
 import org.enso.compiler.data.CompilerConfig;
 import org.enso.home.HomeManager;
 import org.enso.interpreter.Language;
 import org.enso.interpreter.OptionsHelper;
+import org.enso.interpreter.instrument.NotificationHandler;
 import org.enso.interpreter.runtime.builtin.Builtins;
 import org.enso.interpreter.runtime.callable.atom.AtomConstructor;
 import org.enso.interpreter.runtime.scope.TopLevelScope;
-import org.enso.interpreter.runtime.util.ShadowedPackage;
 import org.enso.interpreter.runtime.util.TruffleFileSystem;
 import org.enso.interpreter.util.ScalaConversions;
 import org.enso.pkg.Package;
 import org.enso.pkg.PackageManager;
 import org.enso.pkg.QualifiedName;
+import org.enso.polyglot.LanguageInfo;
 import org.enso.polyglot.RuntimeOptions;
+import scala.jdk.javaapi.CollectionConverters;
 
 /**
  * The language context is the internal state of the language that is associated with each thread in
@@ -42,28 +41,32 @@ public class Context {
 
   private final Language language;
   private final Env environment;
-  private final Compiler compiler;
+  private @CompilationFinal Compiler compiler;
   private final PrintStream out;
   private final PrintStream err;
   private final InputStream in;
   private final BufferedReader inReader;
-  private List<Package<TruffleFile>> packages;
+  private @CompilationFinal PackageRepository packageRepository;
   private @CompilationFinal TopLevelScope topScope;
   private final ThreadManager threadManager;
   private final ResourceManager resourceManager;
   private final boolean isCachingDisabled;
   private final Builtins builtins;
   private final String home;
-  private final List<ShadowedPackage> shadowedPackages;
   private final CompilerConfig compilerConfig;
+  private final NotificationHandler notificationHandler;
+  private final TruffleLogger logger = TruffleLogger.getLogger(LanguageInfo.ID, Context.class);
 
   /**
    * Creates a new Enso context.
    *
    * @param language the language identifier
+   * @param home language home
    * @param environment the execution environment of the {@link TruffleLanguage}
+   * @param notificationHandler a handler for notifications
    */
-  public Context(Language language, String home, Env environment) {
+  public Context(
+      Language language, String home, Env environment, NotificationHandler notificationHandler) {
     this.language = language;
     this.environment = environment;
     this.out = new PrintStream(environment.out());
@@ -75,68 +78,59 @@ public class Context {
     this.isCachingDisabled = environment.getOptions().get(RuntimeOptions.DISABLE_INLINE_CACHES_KEY);
     this.compilerConfig = new CompilerConfig(false, true);
     this.home = home;
-    this.shadowedPackages = new ArrayList<>();
-
-    builtins = new Builtins(this);
-
-    this.compiler =
-        new Compiler(this, builtins, new PackageRepository.Default(this), compilerConfig);
+    this.builtins = new Builtins(this);
+    this.notificationHandler = notificationHandler;
   }
 
   /** Perform expensive initialization logic for the context. */
   public void initialize() {
     TruffleFileSystem fs = new TruffleFileSystem();
-    HashMap<String, Package<TruffleFile>> packageMap = new HashMap<>();
-    packages = new ArrayList<>();
-
-    if (home != null) {
-      HomeManager<TruffleFile> homeManager =
-          new HomeManager<>(environment.getInternalTruffleFile(home), fs);
-      packageMap.putAll(
-          homeManager.loadStdLib().collect(Collectors.toMap((Package::name), Function.identity())));
-    }
-
     PackageManager<TruffleFile> packageManager = new PackageManager<>(fs);
-    List<TruffleFile> packagePaths = OptionsHelper.getPackagesPaths(environment);
 
-    // Add user packages one-by-one, shadowing previously added packages. It assumes that the
-    // standard library packages will not clash. In the future, we should be able to disambiguate
-    // packages that clash.
-    for (var packagePath : packagePaths) {
-      Optional<Package<TruffleFile>> asPackage =
+    Optional<TruffleFile> projectRoot = OptionsHelper.getProjectRoot(environment);
+    Optional<Package<TruffleFile>> projectPackage =
+        projectRoot.flatMap(
+            file -> {
+              var result = packageManager.fromDirectory(projectRoot.get());
+              if (result.isEmpty()) {
+                logger.warning("Could not load the project root package.");
+              }
+              return ScalaConversions.asJava(result);
+            });
+
+    packageRepository =
+        PackageRepository.makeLegacyRepository(
+            RuntimeDistributionManager$.MODULE$, this, builtins, notificationHandler);
+    topScope = new TopLevelScope(builtins, packageRepository);
+    this.compiler = new Compiler(this, builtins, packageRepository, compilerConfig);
+
+    projectPackage.ifPresent(
+        pkg -> packageRepository.registerMainProjectPackage(pkg.libraryName(), pkg));
+
+    List<Package<TruffleFile>> packagesToPreload = new ArrayList<>();
+    // TODO [RW] This preloading mechanism should be replaced by prepending this special path to the
+    // local libraries search paths when switching to the actual edition-based resolution.
+    List<TruffleFile> preloadedPackagePaths = OptionsHelper.getPreloadedPackagesPaths(environment);
+    for (var packagePath : preloadedPackagePaths) {
+      Optional<Package<TruffleFile>> pkgOpt =
           ScalaConversions.asJava(packageManager.fromDirectory(packagePath));
-      if (asPackage.isPresent()) {
-        Package<TruffleFile> pkg = asPackage.get();
-        boolean nameExists = packageMap.containsKey(pkg.name());
-
-        if (nameExists) {
-          shadowedPackages.add(
-              new ShadowedPackage(
-                  packageMap.get(pkg.name()).root().getPath(), pkg.root().getPath(), pkg.name()));
-          packageMap.remove(pkg.name());
-        }
-        packageMap.put(pkg.name(), pkg);
+      if (pkgOpt.isPresent()) {
+        var pkg = pkgOpt.get();
+        packagesToPreload.add(pkg);
+      } else {
+        logger.warning("Could not preload a package.");
       }
     }
 
-    packages.addAll(packageMap.values());
+    // TODO [RW] this is left here temporarily, until the edition system takes over resolving the
+    // std-lib
+    if (home != null) {
+      HomeManager<TruffleFile> homeManager =
+          new HomeManager<>(environment.getInternalTruffleFile(home), fs);
+      homeManager.loadStdLib().forEach(packagesToPreload::add);
+    }
 
-    packages.forEach(
-        pkg -> {
-          List<TruffleFile> classPathItems =
-              ScalaConversions.asJava(pkg.listPolyglotExtensions("java"));
-          classPathItems.forEach(environment::addToHostClassPath);
-        });
-
-    Map<String, Module> knownFiles =
-        packages.stream()
-            .flatMap(
-                p ->
-                    ScalaConversions.asJava(p.listSources()).stream()
-                        .map(srcFile -> new Module(srcFile.qualifiedName(), p, srcFile.file())))
-            .collect(Collectors.toMap(mod -> mod.getName().toString(), mod -> mod));
-
-    topScope = new TopLevelScope(builtins, knownFiles);
+    packageRepository.registerForPreload(CollectionConverters.asScala(packagesToPreload).toSeq());
   }
 
   public TruffleFile getTruffleFile(File file) {
@@ -217,7 +211,7 @@ public class Context {
    */
   public Optional<QualifiedName> getModuleNameForFile(File path) {
     TruffleFile p = getTruffleFile(path);
-    return packages.stream()
+    return ScalaConversions.asJava(packageRepository.getLoadedPackages()).stream()
         .filter(pkg -> p.startsWith(pkg.sourceDir()))
         .map(pkg -> pkg.moduleNameForFile(p))
         .findFirst();
@@ -231,23 +225,7 @@ public class Context {
    * @param newName the new project name
    */
   public void renameProject(String namespace, String oldName, String newName) {
-    renamePackages(namespace, oldName, newName);
-    topScope.renameProjectInModules(namespace, oldName, newName);
-  }
-
-  private void renamePackages(String namespace, String oldName, String newName) {
-    List<Package<TruffleFile>> toChange =
-        packages.stream()
-            .filter(
-                p -> p.config().namespace().equals(namespace) && p.config().name().equals(oldName))
-            .collect(Collectors.toList());
-
-    packages.removeAll(toChange);
-
-    List<Package<TruffleFile>> renamed =
-        toChange.stream().map(p -> p.setPackageName(newName)).collect(Collectors.toList());
-
-    packages.addAll(renamed);
+    packageRepository.renameProject(namespace, oldName, newName);
   }
 
   /**
@@ -294,7 +272,7 @@ public class Context {
     if (file == null) {
       return Optional.empty();
     }
-    return packages.stream()
+    return ScalaConversions.asJava(packageRepository.getLoadedPackages()).stream()
         .filter(pkg -> file.getAbsoluteFile().startsWith(pkg.root().getAbsoluteFile()))
         .findFirst();
   }
@@ -317,7 +295,7 @@ public class Context {
    * @return an object containing the builtin functions
    */
   public Builtins getBuiltins() {
-    return getTopScope().getBuiltins();
+    return this.builtins;
   }
 
   /**
@@ -384,16 +362,6 @@ public class Context {
   /** @return whether inline caches should be disabled for this context. */
   public boolean isCachingDisabled() {
     return isCachingDisabled;
-  }
-
-  /** @return the list of shadowed packages */
-  public List<ShadowedPackage> getShadowedPackages() {
-    return shadowedPackages;
-  }
-
-  /** @return the pre-loaded packages */
-  public List<Package<TruffleFile>> getPackages() {
-    return packages;
   }
 
   /** @return the compiler configuration for this language */