+++ lib/scala/library-manager/src/main/scala/org/enso/librarymanager/local/DefaultLocalLibraryProvider.scala
@@ -1,21 +1,24 @@
 package org.enso.librarymanager.local
 
-import org.enso.distribution.DistributionManager
-import org.enso.editions.LibraryName
+import com.typesafe.scalalogging.Logger
 import org.enso.distribution.FileSystem.PathSyntax
+import org.enso.editions.LibraryName
+import org.enso.logger.masking.MaskedPath
 
 import java.nio.file.{Files, Path}
 import scala.annotation.tailrec
 
 /** A default implementation of [[LocalLibraryProvider]]. */
-class DefaultLocalLibraryProvider(distributionManager: DistributionManager)
+class DefaultLocalLibraryProvider(searchPaths: List[Path])
     extends LocalLibraryProvider {
+
+  private val logger = Logger[DefaultLocalLibraryProvider]
 
   /** @inheritdoc */
   override def findLibrary(libraryName: LibraryName): Option[Path] =
     findLibraryHelper(
       libraryName,
-      distributionManager.paths.localLibrariesSearchPaths.toList
+      searchPaths
     )
 
   /** Searches through the available library paths, checking if any one of them contains the requested library.
@@ -29,9 +32,19 @@ class DefaultLocalLibraryProvider(distributionManager: DistributionManager)
   ): Option[Path] = searchPaths match {
     case head :: tail =>
       val potentialPath = head / libraryName.namespace / libraryName.name
-      if (Files.exists(potentialPath) && Files.isDirectory(potentialPath))
+      if (Files.exists(potentialPath) && Files.isDirectory(potentialPath)) {
+        logger.trace(
+          s"Found a local $libraryName at " +
+          s"[${MaskedPath(potentialPath).applyMasking()}]."
+        )
         Some(potentialPath)
-      else findLibraryHelper(libraryName, tail)
+      } else {
+        logger.trace(
+          s"Local library $libraryName not found at " +
+          s"[${MaskedPath(potentialPath).applyMasking()}]."
+        )
+        findLibraryHelper(libraryName, tail)
+      }
     case Nil => None
   }
 }